requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
PyPDF2==3.0.1
transformers==4.35.2
torch==2.0.1
torchvision==0.15.1
torchaudio==2.0.1
Werkzeug==3.0.1
python-dotenv==1.0.0
requests==2.31.0
